requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
csscompressor==0.9.5
cycler==0.10.0
discord.py==1.6.0
Flask==1.1.2
google-images-download==2.8.0
htmlmin==0.1.12
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
jsmin==2.2.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
more-itertools==8.6.0
multidict==5.1.0
mysql-connector-python==8.0.23
numpy==1.19.5
Pillow==8.1.0
protobuf==3.15.6
pycparser==2.20
pygame==2.0.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
Quandl==3.6.0
requests==2.25.1
scipy==1.6.0
selenium==3.141.0
six==1.15.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.6.3
youtube-dl==2021.3.14